requirements.txt
gunicorn==20.0.4   
flask==2.2.2
music21==8.1.0
numpy==1.19.3
nltk==3.7
tqdm==4.64.1
tensorflow-cpu==2.6.0
protobuf==3.20.0